src/requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
uvicorn==0.27.1
python-multipart==0.0.9
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
PyMuPDF==1.24.9
motor==3.4.0
openai==1.42.0
cohere==5.8.1
qdrant-client==1.11.0
google-generativeai == 0.7.2